# docs/swagger.yaml
basePath: /v1
definitions:
...
  request.CreateRecruiterRequest:
    properties:
      company_contact:
        type: string
      company_description:
        type: string
      company_location:
        type: string
      company_name:
        type: string
      company_website:
        type: string
      social_links:
        type: string
      verified_status:
        type: boolean
    required:
    - company_contact
    - company_description
    - company_location
    - company_name
    - company_website
    - social_links
    - verified_status
    type: object
  request.CreateUsersRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        minLength: 4
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
...
  request.UpdateRecruiterRequest:
    properties:
      company_contact:
        type: string
      company_description:
        type: string
      company_location:
        type: string
      company_name:
        maxLength: 150
        minLength: 3
        type: string
      company_website:
        type: string
      social_links:
        type: string
      verified_status:
        type: boolean
    required:
    - social_links
    type: object
  request.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      role:
        minLength: 4
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: dz-jobs-api-production.up.railway.app
info:
  contact: {}
  description: This is the API documentation for the DZ Jobs portal.
  title: DZ Jobs API
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUsersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new user
      tags:
      - Users
  /admin/users/{user_id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete user
      tags:
      - Users
    get:
      description: Get user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update user
      tags:
      - Users
...
  /recruiters:
    post:
      consumes:
      - multipart/form-data
      description: Create a new recruiter with company logo
      parameters:
      - description: Company Logo
        in: formData
        name: company_logo
        required: true
        type: file
      - description: Recruiter request
        in: body
        name: recruiter
        required: true
        schema:
          $ref: '#/definitions/request.CreateRecruiterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new recruiter
      tags:
      - Recruiters
  /recruiters/{id}:
    delete:
      description: Delete recruiter by ID
      parameters:
      - description: Recruiter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete recruiter
      tags:
      - Recruiters
    get:
      description: Get recruiter details by ID
      parameters:
      - description: Recruiter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get recruiter by ID
      tags:
      - Recruiters
    put:
      consumes:
      - multipart/form-data
      description: Update recruiter details with company logo
      parameters:
      - description: Recruiter ID
        in: path
        name: id
        required: true
        type: string
      - description: Company Logo
        in: formData
        name: company_logo
        required: true
        type: file
      - description: Recruiter request
        in: body
        name: recruiter
        required: true
        schema:
          $ref: '#/definitions/request.UpdateRecruiterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update recruiter
      tags:
      - Recruiters
swagger: "2.0"